requirements.txt
numpy
pandas
matplotlib
keras
tensorflow
ipykernel
seaborn
tqdm
Pillow
pydot
streamlit